[CID-1113] - Augment group deletion to remove group membership from all users

+++ src/commitTest/groovy/com/rackspace/idm/modules/usergroups/service/DefaultUserGroupServiceTest.groovy
@@ -135,18 +135,45 @@ class DefaultUserGroupServiceTest extends RootServiceTest{
         1 * dao.addGroup(group)
     }
 
-    def "deleteGroup: removes object via dao"() {
+    def "deleteGroup: removes object via dao and remove group membership from users"() {
+        mockIdentityUserService(service)
+        def domainId = "domainId"
         UserGroup group = new UserGroup().with {
-            it.domainId = "domainId"
+            it.domainId = domainId
             it.name = "name"
             it.description = "description"
             it
         }
 
-        when:
+        EndUser user1 = new User().with {
+            it.domainId = domainId
+            it.id = "id"
+            it
+        }
+        EndUser user2 = new User().with {
+            it.domainId = domainId
+            it.id = "id2"
+            it
+        }
+
+        when: "Users belong to user group"
         service.deleteGroup(group)
 
         then:
+        1 * identityUserService.getEndUsersInUserGroup(group) >> [user1, user2].asList()
+        1 * identityUserService.checkAndGetUserById(user1.id) >> user1
+        1 * identityUserService.checkAndGetUserById(user2.id) >> user2
+        1 * identityUserService.removeUserGroupFromUser(group, user1)
+        1 * identityUserService.removeUserGroupFromUser(group, user2)
+        1 * dao.deleteGroup(group)
+
+        when: "No users belong to user group"
+        service.deleteGroup(group)
+
+        then:
+        1 * identityUserService.getEndUsersInUserGroup(group) >> [].asList()
+        0 * identityUserService.checkAndGetUserById(_)
+        0 * identityUserService.removeUserGroupFromUser(_, _)
         1 * dao.deleteGroup(group)
     }
 
@@ -161,17 +188,37 @@ class DefaultUserGroupServiceTest extends RootServiceTest{
         1 * dao.getGroupsForDomain(domainId)
     }
 
-    def "getUsersInGroup: gets objects via identityUserService"() {
+    def "getUsersInGroupPaged: gets objects via identityUserService"() {
         given:
         def group = new UserGroup()
         def userSearchCriteria = new UserSearchCriteria(new PaginationParams())
         mockIdentityUserService(service)
 
         when:
-        service.getUsersInGroup(group, userSearchCriteria)
+        service.getUsersInGroupPaged(group, userSearchCriteria)
 
         then:
-        1 * identityUserService.getEndUsersInUserGroup(group, userSearchCriteria)
+        1 * identityUserService.getEndUsersInUserGroupPaged(group, userSearchCriteria)
+    }
+
+    def "getUsersInGroup: gets objects via identityUserService"() {
+        given:
+        def group = new UserGroup()
+        mockIdentityUserService(service)
+
+        when:
+        service.getUsersInGroup(group)
+
+        then:
+        1 * identityUserService.getEndUsersInUserGroup(group)
+    }
+
+    def "getUsersInGroup: null params"() {
+        when:
+        service.getUsersInGroup(null)
+
+        then:
+        thrown(IllegalArgumentException)
     }
 
     def "getGroupByNameForUserInDomain: invalid params"() {
